requirements.txt
requests==2.31.0
gspread==5.7.2
pandas==2.0.1
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
schedule==1.2.0
flask==2.3.2